+++ MOP/src/Helpers/SaveManagerBehaviour.cs
@@ -1,0 +1,58 @@
+﻿using MSCLoader;
+using System;
+using System.Collections;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Text;
+using UnityEngine;
+
+namespace MOP.Helpers
+{
+    internal class SaveManagerBehaviour : MonoBehaviour
+    {
+        public void Run()
+        {
+            if (currentRoutine != null)
+            {
+                StopCoroutine(currentRoutine);
+            }
+            currentRoutine = RemoveReadOnlyAttributesRoutine();
+            StartCoroutine(currentRoutine);
+        }
+
+        private IEnumerator currentRoutine;
+        private IEnumerator RemoveReadOnlyAttributesRoutine()
+        {
+            int retries = 0;
+            while (retries < 5)
+            {
+                try
+                {
+                    RemoveAttribute(SaveManager.SavePath);
+                    RemoveAttribute(SaveManager.ItemsPath);
+                    ModConsole.Log("[MOP] Removed attributes from save file.");
+                    currentRoutine = null;
+                    yield break;
+                }
+                catch
+                {
+                    ModConsole.Log($"[MOP] Retrying removing attributes ({retries + 1}/5)");
+                }
+                retries++;
+                yield return new WaitForSeconds(0.5f);
+            }
+            ModConsole.Error("[MOP] Failed to remove attributes from the save file.");
+
+            currentRoutine = null;
+        }
+
+        static void RemoveAttribute(string filename)
+        {
+            if (File.Exists(filename))
+            {
+                File.SetAttributes(filename, File.GetAttributes(filename) & ~FileAttributes.ReadOnly);
+            }
+        }
+    }
+}